// functions/src/utils/enums/actionType.ts
export enum ActionType {
    Video = "Video",
    Blog = "Blog",
    Quiz = "Quiz",
    Claim = "Claim",
    SocialTwitter = "Social_Twitter",
    SocialTwitterVerify = "Social_Twitter_Verify",
    SocialDiscord = "Social_Discord",
    Notify = "Notify",
    Info = "Info",
    GenerateInviteCode = "Generate_Invite_Code",
    GenerateOGInviteCode = "Generate_OG_Invite_Code",
    ShareInviteCode = "Share_Invite_Code",
    CheckProfile = "CheckProfile",
    VerifyInviteCode = "Verify_Invite_Code",
    DailyReward = "Daily_Reward",
    BuildProfile = "BuildProfile",
    FullName = "updateFullName",
    ProfileImage = "updateProfileImage",
    LinkTwitter = "linkTwitter",
    LinkDiscord = "linkDiscord",
    SubmitForReview = "SubmitForReview",
    FeedbackForm = "FeedbackForm",
    AlphaAirDrop = "Alpha_AirDrop",
    HarborAirdrop = "Harbor_AirDrop",
    VerifyOnChain = "Verify_OnChain",
    ConnectCapxWallet = "Connect_Capx_Wallet"
}