// functions/src/utils/enums/onChainVerifyType.ts
export enum OnChainVerifyType {
    VoteHarbor = "Vote_Harbor",
    MintCMSTVault = "Mint_CMST_Vault",
    MintCMSTStablemint = "Mint_CMST_Stablemint",
    StakeHarbor = "Stake_Harbor",
    LendCMST = "Lend_CMST",
    BorrowWithCMST = "Borrow_With_CMST",
    TradeCSwap = "Trade_CSWAP",
    LiquidityFarmMaster = "Liquidity_Farm_Master",
    LiquidityFarmChild = "Liquidity_Farm_Child"
}