// functions/src/utils/enums/twitterActionType.ts
export enum TwitterActionType {
    UserFollows = "checkIfUserFollows",
    UserLikedTweet = "checkIfUserLikedTweet",
    UserQuoted = "checkIfUserQuoted",
    UserCommented = "checkIfUserCommented",
    UserRetweeted = "checkIfUserRetweeted",
    UserTweet = "checkUserTweet",
    TwitterInfo = "info"
}